+++ provider/pkg/provider/helm_release_test.go
@@ -1,9 +1,24 @@
+// Copyright 2016-2022, Pulumi Corporation.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package provider
 
 import (
+	"testing"
+
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"testing"
 )
 
 func Test_MergeMaps(t *testing.T) {
@@ -29,19 +44,22 @@ func Test_MergeMaps(t *testing.T) {
 
 	for _, test := range []struct {
 		name     string
+		allowNil bool
 		dest     map[string]interface{}
 		src      map[string]interface{}
 		expected map[string]interface{}
 	}{
 		{
 			name:     "Precedence",
+			allowNil: false,
 			dest:     m,
 			src:      override,
 			expected: override, // Expect the override to take precedence
 		},
 		{
-			name: "Merge maps",
-			dest: m,
+			name:     "Merge maps",
+			allowNil: false,
+			dest:     m,
 			src: map[string]interface{}{
 				"a": map[string]interface{}{
 					"b": map[string]interface{}{
@@ -67,8 +85,9 @@ func Test_MergeMaps(t *testing.T) {
 			},
 		},
 		{
-			name: "Dest Has Nil Values",
-			dest: m,
+			name:     "Dest Has Nil Values- disallow nil",
+			allowNil: false,
+			dest:     m,
 			src: map[string]interface{}{
 				"a": map[string]interface{}{
 					"b": map[string]interface{}{
@@ -87,9 +106,33 @@ func Test_MergeMaps(t *testing.T) {
 				},
 			},
 		},
+		{
+			name:     "Dest Has Nil Values- allow nil",
+			allowNil: true,
+			dest:     m,
+			src: map[string]interface{}{
+				"a": map[string]interface{}{
+					"b": map[string]interface{}{
+						"c": interface{}(nil),
+						"e": (*interface{})(nil),
+					},
+				},
+			},
+			expected: map[string]interface{}{
+				"a": map[string]interface{}{
+					"b": map[string]interface{}{
+						"c": interface{}(nil),
+						"e": (*interface{})(nil),
+						"d": []interface{}{
+							"1", "2",
+						},
+					},
+				},
+			},
+		},
 	} {
 		t.Run(test.name, func(t *testing.T) {
-			merged, err := mergeMaps(test.dest, test.src)
+			merged, err := mergeMaps(test.dest, test.src, test.allowNil)
 			require.NoError(t, err)
 			assert.Equal(t, test.expected, merged)
 		})