+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+- Add allowNullValues boolean option to pass Null values through helm configs without having them scrubbed (https://github.com/pulumi/pulumi-kubernetes/issues/2089)
+
 ## 3.22.1 (October 26, 2022)
 
 Note: Enabling SSA mode by default was causing problems for a number of users, so we decided to revert this change.